Findbugs: cannot compare byte with int=255
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/net/trunk@1080620 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/net/telnet/Telnet.java
@@ -794,7 +794,7 @@ class Telnet extends SocketClient
             for (int ii = 0; ii < subn.length; ii++)
             {
                 byte b = (byte) subn[ii];
-                if (b == TelnetCommand.IAC) {
+                if (b == (byte) TelnetCommand.IAC) { // cast is necessary because IAC is outside the signed byte range
                     _output_.write(b); // double any IAC bytes
                 }
                 _output_.write(b);
